engine/requirements.txt
google-api-python-client
python-dotenv
pandas
numpy
matplotlib
seaborn
beautifulsoup4